requirements.txt
streamlit
Pillow
piexif